Fix the namespace Microsoft.EntityFrameworkCore.Sqlite.Storage.Internal.Json
Part of #30306
