+++ src/EFCore.Sqlite.Core/Storage/Json/Internal/SqliteJsonDateTimeReaderWriter.cs
@@ -2,11 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Globalization;
-using System.Text.Encodings.Web;
 using System.Text.Json;
 using Microsoft.EntityFrameworkCore.Storage.Json;
 
-namespace Microsoft.EntityFrameworkCore.Sqlite.Storage.Internal.Json;
+namespace Microsoft.EntityFrameworkCore.Sqlite.Storage.Json.Internal;
 
 /// <summary>
 ///     The Sqlite-specific JsonValueReaderWrite for DateTime. Generates a ISO8601 string representation with a space instead of a T
@@ -18,9 +17,9 @@ namespace Microsoft.EntityFrameworkCore.Sqlite.Storage.Internal.Json;
 ///     any release. You should only use it directly in your code with extreme caution and knowing that
 ///     doing so can result in application failures when updating to a new Entity Framework Core release.
 /// </remarks>
-public sealed class SqliteJsonDateTimeOffsetReaderWriter : JsonValueReaderWriter<DateTimeOffset>
+public sealed class SqliteJsonDateTimeReaderWriter : JsonValueReaderWriter<DateTime>
 {
-    private const string DateTimeOffsetFormatConst = @"{0:yyyy\-MM\-dd HH\:mm\:ss.FFFFFFFzzz}";
+    private const string DateTimeFormatConst = @"{0:yyyy\-MM\-dd HH\:mm\:ss.FFFFFFF}";
 
     /// <summary>
     ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -28,9 +27,9 @@ public sealed class SqliteJsonDateTimeOffsetReaderWriter : JsonValueReaderWriter
     ///     any release. You should only use it directly in your code with extreme caution and knowing that
     ///     doing so can result in application failures when updating to a new Entity Framework Core release.
     /// </summary>
-    public static SqliteJsonDateTimeOffsetReaderWriter Instance { get; } = new();
+    public static SqliteJsonDateTimeReaderWriter Instance { get; } = new();
 
-    private SqliteJsonDateTimeOffsetReaderWriter()
+    private SqliteJsonDateTimeReaderWriter()
     {
     }
 
@@ -40,9 +39,8 @@ public sealed class SqliteJsonDateTimeOffsetReaderWriter : JsonValueReaderWriter
     ///     any release. You should only use it directly in your code with extreme caution and knowing that
     ///     doing so can result in application failures when updating to a new Entity Framework Core release.
     /// </summary>
-    public override DateTimeOffset FromJsonTyped(ref Utf8JsonReaderManager manager, object? existingObject = null)
-        // => manager.CurrentReader.GetDateTimeOffset();
-        => DateTimeOffset.Parse(manager.CurrentReader.GetString()!, CultureInfo.InvariantCulture);
+    public override DateTime FromJsonTyped(ref Utf8JsonReaderManager manager, object? existingObject = null)
+        => DateTime.Parse(manager.CurrentReader.GetString()!, CultureInfo.InvariantCulture);
 
     /// <summary>
     ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -50,10 +48,6 @@ public sealed class SqliteJsonDateTimeOffsetReaderWriter : JsonValueReaderWriter
     ///     any release. You should only use it directly in your code with extreme caution and knowing that
     ///     doing so can result in application failures when updating to a new Entity Framework Core release.
     /// </summary>
-    public override void ToJsonTyped(Utf8JsonWriter writer, DateTimeOffset value)
-        // We use UnsafeRelaxedJsonEscaping to prevent the DateTimeOffset plus (+) sign from getting escaped
-        => writer.WriteStringValue(
-            JsonEncodedText.Encode(
-                string.Format(CultureInfo.InvariantCulture, DateTimeOffsetFormatConst, value),
-                JavaScriptEncoder.UnsafeRelaxedJsonEscaping));
+    public override void ToJsonTyped(Utf8JsonWriter writer, DateTime value)
+        => writer.WriteStringValue(string.Format(CultureInfo.InvariantCulture, DateTimeFormatConst, value));
 }
